fix(types): 修复 ImageToken 序列化递归栈溢出
用 ImageTokenStr 新类型封装纯数据 ImageToken:序列化时自动加密为
字符串,加密不再经 serde 序列化自身,消除 /me 等接口的栈溢出。
DTO 字段与头像上传返回类型改用 ImageTokenStr;nonce seed 纳入
viewer_id 避免 AES-GCM nonce 复用。

+++ types/src/photo/collection.rs
@@ -40,7 +40,7 @@ mod entity {
         pub description: Option<String>,
 
         /// 收藏的照片总数
-        pub photo_count: u64,
+        pub photo_count: i64,
 
         /// 封面照片的文件ID
         pub cover_file_id: Option<String>,
@@ -76,7 +76,7 @@ mod entity {
                 user_id: model.user_id,
                 name: model.name,
                 description: model.description,
-                photo_count: model.photo_count,
+                photo_count: model.photo_count as u64,
                 cover_file_id: model.cover_file_id,
                 cover_photo_id: model.cover_photo_id,
                 created_at: model.created_at,
